+++ AddOns/EditModeExpanded/libs/EditModeExpanded-1.0/EditModeExpanded-1.0.lua
@@ -3,20 +3,24 @@
 --
 
 local CURRENT_BUILD = "10.0.2"
-local MAJOR, MINOR = "EditModeExpanded-1.0", 23
+local MAJOR, MINOR = "EditModeExpanded-1.0", 25
 local lib = LibStub:NewLibrary(MAJOR, MINOR)
 if not lib then return end
 
--- the internal frames provided by Blizzard go up to index 12. They reference an Enum.
-local index = 13
+-- the internal frames provided by Blizzard go up to index 12. They reference Enum.EditModeSystem, which starts from index 0
+local STARTING_INDEX = 13
+local index = STARTING_INDEX
 local frames = {}
 local baseFramesDB = {} -- the base db that includes all profiles inside
 local framesDB = {} -- the currently selected db inside the profile
 local framesDialogs = {}
 local framesDialogsKeys = {}
+local existingFrames = {} -- frames already part of Edit Mode where we are adding more options
+local firstCheckButtonPlaced = false
 
 local ENUM_EDITMODEACTIONBARSETTING_HIDEABLE = 10 -- Enum.EditModeActionBarSetting.Hideable = 10
 local ENUM_EDITMODEACTIONBARSETTING_MINIMAPPINNED = 11
+local ENUM_EDITMODEACTIONBARSETTING_CUSTOM = 12
 
 -- run OnLoad the first time RegisterFrame is called by an addon
 local f = {}
@@ -47,6 +51,13 @@ end
 local pinToMinimap
 local unpinFromMinimap
 local getOffsetXY
+
+local function getSystemID(frame)
+    if frame.system < STARTING_INDEX then
+        return frame.EMESystemID
+    end
+    return frame.system
+end
 
 --
 -- Code to deal with splitting the Main Menu Bar from the Backpack bar
@@ -156,6 +167,38 @@ function lib:RegisterFrame(frame, name, db, anchorTo, anchorPoint)
     
     local baseDB = db
     
+    -- If the frame was already registered (perhaps by another addon that uses this library), don't register it again
+    for _, f in ipairs(frames) do
+        if (frame == f) or ((frame == MicroButtonAndBagsBar) and (f == MicroButtonAndBagsBarMovable)) then
+            if (not framesDB[f.system].x) and (not framesDB[f.system].y) then
+                -- import new db settings if there are none saved in the existing db
+                framesDB[f.system].x = db.x
+                framesDB[f.system].y = db.y
+                local x, y = getOffsetXY(frame, db.x, db.y)
+                f:SetPoint(anchorPoint, anchorTo, anchorPoint, x, y)
+            end
+            return
+        end
+    end
+    
+    local systemID = frame.system
+    
+    -- frame is an existing Edit Mode frame by Blizzard, handle it differently
+    if systemID then
+        if existingFrames[frame:GetName()] then return end
+        existingFrames[frame:GetName()] = true
+        
+        systemID = index
+        index = index + 1
+        frame.EMESystemID = systemID
+        baseFramesDB[systemID] = baseDB
+        framesDB[systemID] = db
+        hooksecurefunc(frame, "SelectSystem", function()
+            EditModeExpandedSystemSettingsDialog:AttachToSystemFrame(frame)
+        end)
+        return
+    end
+    
     if profilesInitialised then
         local layoutInfo = EditModeManagerFrame:GetActiveLayoutInfo()
         local profileName = layoutInfo.layoutType.."-"..layoutInfo.layoutName
@@ -182,21 +225,7 @@ function lib:RegisterFrame(frame, name, db, anchorTo, anchorPoint)
         
         db = db.profiles[profileName]
     end
-    
-    -- If the frame was already registered (perhaps by another addon that uses this library), don't register it again
-    for _, f in ipairs(frames) do
-        if (frame == f) or ((frame == MicroButtonAndBagsBar) and (f == MicroButtonAndBagsBarMovable)) then
-            if (not framesDB[f.system].x) and (not framesDB[f.system].y) then
-                -- import new db settings if there are none saved in the existing db
-                framesDB[f.system].x = db.x
-                framesDB[f.system].y = db.y
-                local x, y = getOffsetXY(frame, db.x, db.y)
-                f:SetPoint(anchorPoint, anchorTo, anchorPoint, x, y)
-            end
-            return
-        end
-    end
-    
+
     if not IsAddOnLoaded("Dominos") and frame == MicroButtonAndBagsBar then -- 自行加入達美樂的相容性
         frame = duplicateMicroButtonAndBagsBar(db)
         frame.EMEanchorTo = anchorTo
@@ -367,11 +396,19 @@ function lib:RegisterFrame(frame, name, db, anchorTo, anchorPoint)
     checkButtonFrame.Text:SetFontObject(GameFontHighlightMedium)
     
     checkButtonFrame.index = frame.system
-    if frame.system == 13 then
+    if not firstCheckButtonPlaced then
+        firstCheckButtonPlaced = true
         checkButtonFrame:SetPoint("TOPLEFT", EditModeManagerExpandedFrame.AccountSettings, "TOPLEFT", 20, 0)
         checkButtonFrame:SetSize(32, 32)
     else
-        checkButtonFrame:SetPoint("TOPLEFT", EditModeManagerExpandedFrame.AccountSettings[frame.system - 1], "BOTTOMLEFT", 0, 10)
+        -- some system IDs may be existing edit mode frames which were not assigned a checkbox
+        local previousSystemID = frame.system - 1
+        local i = 1
+        while not EditModeManagerExpandedFrame.AccountSettings[previousSystemID] do
+            i = i + 1
+            previousSystemID = frame.system - i
+        end
+        checkButtonFrame:SetPoint("TOPLEFT", EditModeManagerExpandedFrame.AccountSettings[previousSystemID], "BOTTOMLEFT", 0, 10)
         checkButtonFrame:SetSize(32, 32)
     end
     
@@ -380,7 +417,7 @@ function lib:RegisterFrame(frame, name, db, anchorTo, anchorPoint)
     
     function frame:GetSettingValue(setting, useRawValue)
         local db = framesDB[frame.system]
-    	if (not self:IsInitialized()) or (not db.settings) or (not db.settings[settings]) then
+    	if (not self:IsInitialized()) or (not db.settings) or (not db.settings[setting]) then
     		return 0;
     	end
     	return db.settings[setting]
@@ -456,7 +493,11 @@ end
 
 -- call this if the frame needs to be moved back into position at some point after ADDON_LOADED
 function lib:RepositionFrame(frame)
-    local db = framesDB[frame.system]
+    assert(type(frame) == "table")
+    assert(frame.system > 12)
+    
+    local systemID = getSystemID(frame)
+    local db = framesDB[systemID]
 
     frame:ClearAllPoints()
     
@@ -477,8 +518,10 @@ function lib:RepositionFrame(frame)
 end
 
 -- Call this to add a slider to the frames dialog box, allowing is to be resized using frame:SetScale
--- param1: custom system frame
+-- param1: an edit mode registered frame, either one already registered by Blizz, or a custom one you have registered with lib:RegisterFrame
 function lib:RegisterResizable(frame)
+    local systemID = getSystemID(frame)
+    
     if not framesDialogs[frame.system] then framesDialogs[frame.system] = {} end
     if framesDialogsKeys[frame.system] and framesDialogsKeys[frame.system][Enum.EditModeUnitFrameSetting.FrameSize] then return end
     if not framesDialogsKeys[frame.system] then framesDialogsKeys[frame.system] = {} end
@@ -497,18 +540,26 @@ function lib:RegisterResizable(frame)
 end
  
 -- Call this to add a checkbox to the frames dialog box, allowing the frame to be permanently hidden outside of Edit Mode
--- param1: custom system frame
+-- param1: an edit mode registered frame, either one already registered by Blizz, or a custom one you have registered with lib:RegisterFrame
 function lib:RegisterHideable(frame)
-    if not framesDialogs[frame.system] then framesDialogs[frame.system] = {} end
-    if framesDialogsKeys[frame.system] and framesDialogsKeys[frame.system][ENUM_EDITMODEACTIONBARSETTING_HIDEABLE] then return end
-    if not framesDialogsKeys[frame.system] then framesDialogsKeys[frame.system] = {} end
-    framesDialogsKeys[frame.system][ENUM_EDITMODEACTIONBARSETTING_HIDEABLE] = true
-    table.insert(framesDialogs[frame.system],
+    local systemID = getSystemID(frame)
+    
+    if not framesDialogs[systemID] then framesDialogs[systemID] = {} end
+    if framesDialogsKeys[systemID] and framesDialogsKeys[systemID][ENUM_EDITMODEACTIONBARSETTING_HIDEABLE] then return end
+    if not framesDialogsKeys[systemID] then framesDialogsKeys[systemID] = {} end
+    framesDialogsKeys[systemID][ENUM_EDITMODEACTIONBARSETTING_HIDEABLE] = true
+    table.insert(framesDialogs[systemID],
         {
             setting = ENUM_EDITMODEACTIONBARSETTING_HIDEABLE,
             name = "隱藏",
             type = Enum.EditModeSettingDisplayType.Checkbox,
     })
+end
+
+function lib:IsFrameMarkedHidden(frame)
+	local systemID = getSystemID(frame)
+    if not framesDB[systemID].settings then framesDB[systemID].settings = {} end
+    return framesDB[systemID].settings[ENUM_EDITMODEACTIONBARSETTING_HIDEABLE] == 1
 end
 
 -- implemented further down the file
@@ -517,16 +568,52 @@ end
 
 -- a simple check of "has this frame been registered with EME" - maybe you want to test if another addon registered it already?
 function lib:IsRegistered(frame)
-    for _, f in pairs(frames) do
-        if frame == f then return true end
-    end
-    return false
+    local systemID = getSystemID(frame)
+    if not systemID then return false end
+    
+    if not framesDB[systemID] then return false end
 end
 
 -- Is the Expanded frame checkbox checked for this frame?
 function lib:IsFrameEnabled(frame)
     local db = framesDB[frame.system]
     return db.enabled
+end
+
+local customCheckboxCallDuringProfileInit = {}
+-- call this to register a custom checkbox where you're providing the handler
+-- limit one per frame
+function lib:RegisterCustomCheckbox(frame, name, onChecked, onUnchecked)
+    local systemID = getSystemID(frame)
+    
+    if not framesDialogs[systemID] then framesDialogs[systemID] = {} end
+    if framesDialogsKeys[systemID] and framesDialogsKeys[systemID][ENUM_EDITMODEACTIONBARSETTING_CUSTOM] then return end
+    if not framesDialogsKeys[systemID] then framesDialogsKeys[systemID] = {} end
+    framesDialogsKeys[systemID][ENUM_EDITMODEACTIONBARSETTING_CUSTOM] = true
+    table.insert(framesDialogs[systemID],
+        {
+            setting = ENUM_EDITMODEACTIONBARSETTING_CUSTOM,
+            name = name,
+            type = Enum.EditModeSettingDisplayType.Checkbox,
+            onChecked = onChecked,
+            onUnchecked = onUnchecked,
+    })
+    
+    local function callLater()
+        local db = framesDB[systemID]
+        if not db.settings then db.settings = {} end
+        if db.settings[ENUM_EDITMODEACTIONBARSETTING_CUSTOM] == 1 then
+            onChecked()
+        else
+            onUnchecked()
+        end
+    end
+    
+    if profilesInitialised then
+        callLater()
+    else
+        table.insert(customCheckboxCallDuringProfileInit, callLater)
+    end
 end
 
 --
@@ -629,17 +716,18 @@ hooksecurefunc(EditModeManagerFrame, "ExitEditMode", function()
 end)
 
 hooksecurefunc(EditModeManagerFrame, "SelectSystem", function(self, systemFrame)
-    local wasCustom
     for _, frame in ipairs(frames) do
         if systemFrame ~= frame then
             frame:HighlightSystem()
-        else
-            wasCustom = true
         end
     end
-    if not wasCustom then
+    
+    local systemID = getSystemID(systemFrame)
+    
+    if (not systemID) or (not framesDB[systemID]) then
         EditModeExpandedSystemSettingsDialog:Hide()
-    end            
+        return
+    end
 end)
 
 --
@@ -745,8 +833,9 @@ hooksecurefunc(f, "OnLoad", function()
             local draggingSlider = self:ReleaseNonDraggingSliders();
     
     		local settingsToSetup = {};
-    
-    		local systemSettingDisplayInfo = GetSystemSettingDisplayInfo(framesDialogs[self.attachedToSystem.system]);
+            local systemID = getSystemID(self.attachedToSystem)
+    		
+            local systemSettingDisplayInfo = GetSystemSettingDisplayInfo(framesDialogs[systemID]);
     		for index, displayInfo in ipairs(systemSettingDisplayInfo) do
   				local settingPool = self:GetSettingPool(displayInfo.type);
 				if settingPool then
@@ -764,9 +853,9 @@ hooksecurefunc(f, "OnLoad", function()
                     
                     local settingName = (self.attachedToSystem:UseSettingAltName(displayInfo.setting) and displayInfo.altName) and displayInfo.altName or displayInfo.name;
   					local updatedDisplayInfo = self.attachedToSystem:UpdateDisplayInfoOptions(displayInfo);
-                    if not framesDB[self.attachedToSystem.system].settings then framesDB[self.attachedToSystem.system].settings = {} end
+                    if not framesDB[systemID].settings then framesDB[systemID].settings = {} end
   					
-                    local savedValue = framesDB[self.attachedToSystem.system].settings[updatedDisplayInfo.setting]
+                    local savedValue = framesDB[systemID].settings[updatedDisplayInfo.setting]
                     
                     if displayInfo.setting == Enum.EditModeUnitFrameSetting.FrameSize then
                         savedValue = savedValue or 100
@@ -785,38 +874,53 @@ hooksecurefunc(f, "OnLoad", function()
                     end
                     
                     if displayInfo.setting == ENUM_EDITMODEACTIONBARSETTING_HIDEABLE then
-                        savedValue = framesDB[EditModeExpandedSystemSettingsDialog.attachedToSystem.system].settings[displayInfo.setting]
+                        savedValue = framesDB[systemID].settings[displayInfo.setting]
                         if savedValue == nil then savedValue = 0 end
                         settingFrame.Button:SetChecked(savedValue)
                         settingFrame.Button:SetScript("OnClick", function()
                             if settingFrame.Button:GetChecked() then
-                                framesDB[EditModeExpandedSystemSettingsDialog.attachedToSystem.system].settings[displayInfo.setting] = 1
+                                framesDB[systemID].settings[displayInfo.setting] = 1
                                 EditModeExpandedSystemSettingsDialog.attachedToSystem:Hide()
-                                wasVisible[EditModeExpandedSystemSettingsDialog.attachedToSystem.system] = false
+                                wasVisible[systemID] = false
                             else
-                                framesDB[EditModeExpandedSystemSettingsDialog.attachedToSystem.system].settings[displayInfo.setting] = 0
+                                framesDB[systemID].settings[displayInfo.setting] = 0
                                 EditModeExpandedSystemSettingsDialog.attachedToSystem:Show()
-                                wasVisible[EditModeExpandedSystemSettingsDialog.attachedToSystem.system] = true
+                                wasVisible[systemID] = true
                             end
                         end)
                     end
                     
                     if displayInfo.setting == ENUM_EDITMODEACTIONBARSETTING_MINIMAPPINNED then
-                        savedValue = framesDB[EditModeExpandedSystemSettingsDialog.attachedToSystem.system].settings[displayInfo.setting]
+                        savedValue = framesDB[systemID].settings[displayInfo.setting]
                         if savedValue == nil then savedValue = 0 end
                         settingFrame.Button:SetChecked(savedValue)
                         settingFrame.Button:SetScript("OnClick", function()
                             if settingFrame.Button:GetChecked() then
-                                framesDB[EditModeExpandedSystemSettingsDialog.attachedToSystem.system].settings[displayInfo.setting] = 1
+                                framesDB[systemID].settings[displayInfo.setting] = 1
                                 pinToMinimap(EditModeExpandedSystemSettingsDialog.attachedToSystem)
                             else
-                                framesDB[EditModeExpandedSystemSettingsDialog.attachedToSystem.system].settings[displayInfo.setting] = 0
+                                framesDB[systemID].settings[displayInfo.setting] = 0
                                 unpinFromMinimap(EditModeExpandedSystemSettingsDialog.attachedToSystem)
                             end
                         end)
                     end
                     
-  					settingsToSetup[settingFrame] = { displayInfo = updatedDisplayInfo, currentValue = savedValue, settingName = settingName },
+                    if displayInfo.setting == ENUM_EDITMODEACTIONBARSETTING_CUSTOM then
+                        savedValue = framesDB[systemID].settings[displayInfo.setting]
+                        if savedValue == nil then savedValue = 0 end
+                        settingFrame.Button:SetChecked(savedValue)
+                        settingFrame.Button:SetScript("OnClick", function()
+                            if settingFrame.Button:GetChecked() then
+                                framesDB[systemID].settings[displayInfo.setting] = 1
+                                displayInfo.onChecked()
+                            else
+                                framesDB[systemID].settings[displayInfo.setting] = 0
+                                displayInfo.onUnchecked()
+                            end
+                        end)
+                    end
+                    
+  					settingsToSetup[settingFrame] = { displayInfo = updatedDisplayInfo, currentValue = savedValue, settingName = settingName }
   					settingFrame:Show();
   				end
     		end
@@ -842,7 +946,7 @@ hooksecurefunc(f, "OnLoad", function()
     function EditModeExpandedSystemSettingsDialog:OnSettingValueChanged(setting, value)
         local attachedToSystem = self.attachedToSystem
     	if attachedToSystem then
-            local db = framesDB[attachedToSystem.system]
+            local db = framesDB[getSystemID(attachedToSystem)]
             if not db.settings then db.settings = {} end
             db.settings[setting] = value
             if setting == Enum.EditModeUnitFrameSetting.FrameSize then
@@ -968,6 +1072,10 @@ do
             if EditModeManagerFrame.editModeActive and frame:IsShown() then
                 frame:HighlightSystem()
             end
+        end
+        
+        for _, func in pairs(customCheckboxCallDuringProfileInit) do
+            func()
         end
     end)
 end