+++ AddOns/EditModeExpanded/CHANGELOG.md
@@ -1,7 +1,19 @@
 # Edit Mode Expanded
 
-## [10.0-023](https://github.com/teelolws/EditModeExpanded/tree/10.0-023) (2022-11-18)
-[Full Changelog](https://github.com/teelolws/EditModeExpanded/compare/10.0-022...10.0-023) [Previous Releases](https://github.com/teelolws/EditModeExpanded/releases)
+## [10.0-024](https://github.com/teelolws/EditModeExpanded/tree/10.0-024) (2022-11-20)
+[Full Changelog](https://github.com/teelolws/EditModeExpanded/compare/10.0-023...10.0-024) [Previous Releases](https://github.com/teelolws/EditModeExpanded/releases)
 
-- Version 23  
-- Revert "Fix profile handling bug with minimap icons."  
+- Add option to hide the Talking Head Frame.  
+- Fix issue with raid frames being moved slightly  
+    https://github.com/teelolws/EditModeExpanded/issues/36#issuecomment-1320939411  
+- Add option for compact raid frames  
+    Fix issue with selecting raid frames, then selecting another frame in its same "family", but the extra dialog box did not go away  
+- Fix Lua error  
+    Closes https://github.com/teelolws/EditModeExpanded/issues/38  
+- Version 24  
+- Update EditModeExpanded-1.0.lua  
+    https://github.com/teelolws/EditModeExpanded/issues/36  
+- Attempt to add toggle option to hide the Compact Raid Frame Manager  
+    The option is added to the CompactRaidFrameContainer dialog.  
+    See: https://github.com/teelolws/EditModeExpanded/issues/36  
+- Add support for a custom checkbox in the custom dialog box